requirements.txt
# Core framework
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.4.0
requests>=2.25.0
pyyaml>=5.4.0

# Agent + Streamlit UI (optional — required only for `streamlit run app.py`)
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-docx>=0.8.11
